refactor: Http Response Formatter 분리
